+++ bin/extract_to_arb.dart
@@ -63,6 +63,7 @@ main(List<String> args) {
   if (locale != null) {
     allMessages["@@locale"] = locale;
   }
+  allMessages["@@last_modified"] = new DateTime.now().toIso8601String();
   for (var arg in args.where((x) => x.contains(".dart"))) {
     var messages = extraction.parseFile(new File(arg), transformer);
     messages.forEach((k, v) => allMessages.addAll(toARB(v)));